Editing topics in place Instead of allowing immediate editing of each topic use a manager with an undo action.
