+++ app/components/detail-topic-list.js
@@ -1,0 +1,7 @@
+import Ember from 'ember';
+
+export default Ember.Component.extend({
+  topics: [],
+  sortBy: ['title'],
+  sortedTopics: Ember.computed.sort('topics', 'sortBy'),
+});
