+++ app/templates/components/detail-topic-list.hbs
@@ -1,0 +1,5 @@
+<ul class='inline-list'>
+  {{#each topic in sortedTopics}}
+    <li>{{topic.title}}</li>
+  {{/each}}
+</ul>
