+++ app/translations/en.js
@@ -195,6 +195,7 @@ var courses = {
   'missingCohortMessage': 'Please add at least one cohort to this course.',
   'confirmRemove': 'Are you sure you want to delete this course, with {{publishedOfferingCount}} published offerings? This action will remove all sessions and offerings for this course, and cannot be undone.',
   'meshSearchPlaceholder': 'Search MeSH',
+  'topicsManageTitle': 'Manage Topics',
 };
 translations.courses = courses;
